+++ app/store/stashLinkHandler.ts
@@ -5,6 +5,7 @@ import { select as selectEntry } from '../actions/currentEntry';
 import { expand } from '../actions/treeState';
 import {Dispatch, Thunk} from '../actions/types/index';
 import {toastr} from 'react-redux-toastr';
+import StashLink from '../domain/StashLink';
 import {onceAfterAction} from './eventMiddleware';
 import * as Repository from '../actions/repository';
 import {hierarchy} from '../utils/repository';
@@ -20,8 +21,8 @@ export function openStashLink(link: string): Thunk<Promise<void>> {
     try {
       console.log('Opening link: ', link);
 
-      const ptr = EntryPtr.fromHref(link);
-      const hier = hierarchy(repository.nodes, ptr.nodeId);
+      const stashLink = StashLink.parse(link);
+      const hier = hierarchy(repository.nodes, stashLink.nodeId);
       if (!hier.length) {
         toastr.error('Invalid link', 'Folder does not exist');
         return;
@@ -30,10 +31,14 @@ export function openStashLink(link: string): Thunk<Promise<void>> {
       for (const node of hier.slice(0, -1)) {
         await dispatch(expand(node.id));
       }
-      await dispatch(selectNode(ptr.nodeId));
-      // TODO the delay is required because selectNode asynchronously clears the selection, and should of course be refactored ;)
-      await new Promise(resolve => setTimeout(resolve, 250));
-      await dispatch(selectEntry(ptr));
+      await dispatch(selectNode(stashLink.nodeId));
+      if (stashLink.isEntry()) {
+        // TODO the delay is required because selectNode asynchronously clears the selection, and should of course be refactored ;)
+        await new Promise(resolve => setTimeout(resolve, 250));
+        if (!await dispatch(selectEntry(stashLink.toEntryPtr()))) {
+          toastr.error('Invalid link', 'Entry not found in folder');
+        }
+      }
     } catch (e) {
       // failed to parse, ignore
       console.error(e);
@@ -42,9 +47,9 @@ export function openStashLink(link: string): Thunk<Promise<void>> {
   };
 }
 
-export function copyStashLink(ptr: EntryPtr) {
-  clipboard.writeText(ptr.toHref());
-  toastr.success('', 'Stash link to entry copied', { timeOut: 2000 });
+export function copyStashLink(ptrOrNodeId: EntryPtr | string) {
+  clipboard.writeText(new StashLink(ptrOrNodeId).toUri());
+  toastr.success('', `Stash link to ${typeof ptrOrNodeId === 'string' ? 'folder' : 'entry'} copied`, {timeOut: 2000});
 }
 
 export default function(dispatch: Dispatch) {